Better scroll handling for trackpads in map view
The wxMouseEvent has two Properties for "Wheels" the delta and the
rotation. Before SAGA just checks if the rotation is negativ or positive
and than zooms the map in or out. On a classical Mouse the Rotation is
the same as the delta. So it is fine to assume to scroll one line.
On a trackpad on the other hand the rotation is more granular. Maybe the
user is not that precise between the sampling and has in a downward
gesture some upwards samples.
The wx Documentation suggest two methods: Accumulate or partial. The
existing method has a fix zoom in and zoom out value so do the
accumulate variant.
Now it collects all wheel inputs and zooms if the delta threshold is
exceeded.