Polymorphism: Use CSG_Shape::Assign with COPY_ATTR
Splitting a Polygon in the GUI with Line via Context menu results in a
lost of Data on one side of the split and a two overlaying polygons on
the other side.
The code was previously intended to use the CSG_Taple_Record::Assign
function via casting to copy the attributes to the new polygon. But used
was CSG_Shape::Assign with SHAPE_COPY as Mode. This results in a
assignment of the geometry.
Now it uses SHAPE_COPY_ATTR and only assigns the Table_Record.