Removed NoData-Speciality for m_Edit_Attributes
m_Edit_Attributes has its own NoData value as CSG_DataObject. If the
NoData-Value of the CSG_Grid is different, m_Edit_Attributes uses its
own. This leads to confusion when users assume to find the correct
values here.
Removing this speciality streamlines the code and saves us from
constantly changing the NoData value depending on the NoData value of
the corresponding DataObject especially in the polymorphic use of this
table.