table_calculus/Table_Calculator: Force refresh of field choice
If the user performs a calculation, manipulates the table in the GUI,
i.e. adds a new field and then reopens the tool, only the
On_Parameters_Enable() is called by Parameter Control. This results in
the field selection not being updated. Now the On_Parameters_Enable()
calls the On_Parameter_Changed() and thus updates the choice.