cmake/modules/FindPostgres.cmake: Use CMake version, if present
On Unix-like systems, the cmake configuration gives the following error:

CMake Warning (dev) at cmake/modules/FindPostgres.cmake:58 (EXEC_PROGRAM):
  Policy CMP0153 is not set: The exec_program command should not be called.
  Run "cmake --help-policy CMP0153" for policy details.  Use the cmake_policy
  command to set the policy and suppress this warning.

  Use execute_process() instead.
Call Stack (most recent call first):
  src/tools/db/db_pgsql/CMakeLists.txt:16 (find_package)
This warning is for project developers.  Use -Wno-dev to suppress it.

This can be avoided by using a more up-to-date version of FindPostgreSQL.cmake
from the QGIS projekt. If the new module finds the FindPostgreSQL version of
the cmake project, pg_config is called from there, the corresponding variables
are set and FindPostgreSQL.cmake is then exited.

If no FindPostgreSQL.cmake was found in the cmake project, SAGA behaves as
before. This seems to work at least for BSD-Unixes and Windows 11.